+++ extensions/better-uptime/src/add-incident.tsx
@@ -1,0 +1,35 @@
+import { Form, ActionPanel, Action, showToast, Toast, getPreferenceValues, popToRoot } from "@raycast/api";
+import axios from "axios";
+import { useRef, useState } from "react";
+import { Preferences } from "./interface";
+
+export default function Command(): JSX.Element {
+  const preferences = getPreferenceValues<Preferences>();
+
+  async function handleSubmit(item: any) {
+    axios
+      .post("https://betteruptime.com/api/v2/incidents", item, {
+        headers: { Authorization: `Bearer ${preferences.apiKey}` },
+      })
+      .then(() => {
+        popToRoot();
+
+        showToast({ title: "Success", message: "Successfully added incident" });
+      })
+      .catch((error) => {
+        showToast(Toast.Style.Failure, "An error occurred", error.response.data.errors);
+      });
+  }
+
+  return (
+    <Form
+      actions={
+        <ActionPanel>
+          <Action.SubmitForm title="Add Incident" onSubmit={handleSubmit} />
+        </ActionPanel>
+      }
+    >
+      <Form.TextField id="url" title="URL" placeholder="https://raycast.com" />
+    </Form>
+  );
+}