+++ extensions/better-uptime/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Better Uptime Changelog
 
+## [Update] - {PR_MERGE_DATE}
+
+- Add the ability to create monitors from within Raycast
+
 ## [Added Better Uptime] - 2022-06-17
 
 - Initial version code